+++ apps/invoicing/management/commands/monthly_invoices.py
@@ -6,19 +6,27 @@ from django.utils import timezone
 from django.conf import settings
 from django.utils.formats import date_format
 
-from KlimaKar.email import mail_managers
+from KlimaKar.email import mail_managers, get_email_message
 from apps.invoicing.models import SaleInvoice
+from apps.settings.models import SiteSettings
 
 
 class Command(BaseCommand):
-    help = "Generate monthly pdf invoices and send email"
+    help = "Generate pdf invoices from previous month and send email"
 
     def add_arguments(self, parser):
         parser.add_argument("date")
 
     def handle(self, **options):
+        config = SiteSettings.load()
+        if not config.SEND_SALE_INVOICE:
+            print("Sending email is disabled in settings.")
+            return
+        if not config.SEND_SALE_INVOICE_EMAILS:
+            print("Email addresses list is empty.")
+            return
         input_date = date_parser.parse(options["date"]).date()
-        date_from = input_date.replace(day=1)
+        date_from = input_date.replace(day=1) - relativedelta(months=1)
         date_to = date_from + relativedelta(months=1) - relativedelta(days=1)
         month_name = date_format(date_from, "F").lower()
         invoices = (
@@ -27,13 +35,19 @@ class Command(BaseCommand):
             .order_by("issue_date", "number_year", "number_value")
         )
 
-        mail_managers(
-            f"faktury {month_name} {date_from.year}",
-            "\n".join([str(i) for i in invoices]),
-            attachment={
-                "filename": f"klimakar_faktury_{date_from.year}{date_from.month}.pdf",
-                "file": invoices.generate_pdf(),
-                "mime": "application/pdf",
-            },
-            fail_silently=False,
+        mail = get_email_message(
+            subject=config.SEND_SALE_INVOICE_TITLE.format(
+                month_name=month_name, month=date_from.month, year=date_from.year
+            ),
+            body=config.SEND_SALE_INVOICE_BODY.format(
+                month_name=month_name, month=date_from.month, year=date_from.year
+            ),
+            to=config.SEND_SALE_INVOICE_EMAILS.split(","),
+            bcc=True,
         )
+        mail.attach(
+            f"klimakar_faktury_{date_from.year}{date_from.month}.pdf",
+            invoices.generate_pdf(),
+            "application/pdf",
+        )
+        mail.send(fail_silently=False)